Rename test packages to same name as source packages. Return IMport from updateImportInProgress().
