--- src/test/java/com/deloitte/elrr/datasych/test/scheduler/package-info.java
@@ -1,1 +1,0 @@
-package com.deloitte.elrr.datasych.test.scheduler;
