+++ src/test/java/com/deloitte/elrr/datasync/entity/AuditableTest.java
@@ -1,12 +1,12 @@
 /** */
-package com.deloitte.elrr.test.datasync.test.entity;
+package com.deloitte.elrr.datasync.entity;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.deloitte.elrr.datasync.entity.Auditable;
-import com.deloitte.elrr.test.datasync.test.util.ValueObjectTestUtility;
+import com.deloitte.elrr.test.datasync.util.ValueObjectTestUtility;
 
 /**
  * @author mnelakurti
