--- src/test/java/com/deloitte/elrr/datasync/test/service/package-info.java
@@ -1,1 +1,0 @@
-package com.deloitte.elrr.datasync.test.service;
